Initial Commit
Initial commit with first working version of MatrixToArff convert. This
converts a matrix created and saved using default options in Octave. The
matrix cannot be staggered. If the project is exported to a jar it can
be run as a standalone in whatever folder you would like from the
command line.